Extract CoreData managed model into separate files.
